+++ lib/generators/crud/controller/templates/authorization.rb
@@ -1,0 +1,4 @@
+<% module_namespacing do -%>
+class <%= class_name %>Authorization < Crud::Authorization::Default
+end
+<% end -%>