+++ spec/dummy/app/authorizations/ar/groups_authorization.rb
@@ -1,0 +1,2 @@
+class Ar::GroupsAuthorization < GroupsAuthorization
+end