+++ spec/dummy/app/authorizations/ar/users_authorization.rb
@@ -1,0 +1,2 @@
+class Ar::UsersAuthorization < UsersAuthorization
+end