+++ spec/dummy/app/authorizations/mongo/groups_authorization.rb
@@ -1,0 +1,2 @@
+class Mongo::GroupsAuthorization < GroupsAuthorization
+end