+++ spec/dummy/app/authorizations/mongo/users_authorization.rb
@@ -1,0 +1,2 @@
+class Mongo::UsersAuthorization < UsersAuthorization
+end