+++ spec/dummy/app/controllers/ar/groups_controller.rb
@@ -20,23 +20,4 @@ class Ar::GroupsController < Crud::ApplicationController
       false
     end
   end
-
-  class Authorization < Crud::Authorization::Default
-    def show?(group)
-      admin? || group.authorized?(current_user, :read)
-    end
-
-    def create?(group)
-      current_user
-    end
-
-    def manage?(group)
-      admin? || group.authorized?(current_user, :manage)
-    end
-
-    private
-    def admin?
-      current_user.try(:is_admin)
-    end
-  end
 end