+++ spec/dummy/app/controllers/ar/users_controller.rb
@@ -31,14 +31,4 @@ class Ar::UsersController < Crud::ApplicationController
     reverse_order = order == :asc ? :desc : :asc
     "birth_date #{reverse_order}"
   end
-
-  class Authorization < Crud::Authorization::Default
-    def manage?(user)
-      current_user.try(:is_admin) || user == current_user
-    end
-
-    def destroy?(user)
-      manage?(user) && user != current_user
-    end
-  end
 end