+++ spec/dummy/app/controllers/mongo/misc_belongings_controller.rb
@@ -11,7 +11,4 @@ class Mongo::MiscBelongingsController < Crud::ApplicationController
   def columns_for_index
     model_columns
   end
-
-  class Authorization < Crud::Authorization::Default
-  end
 end